backend/requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# WSGI server
gunicorn==21.2.0

# Static file serving
whitenoise==6.6.0

# Environment variables
python-dotenv==1.0.0

# Template engine
Jinja2==3.1.2

# HTTP client
requests==2.31.0

# System monitoring
psutil==5.9.6

# Date/time utilities
python-dateutil==2.8.2

# JSON Web Tokens (optional)
PyJWT==2.8.0

# Database drivers
# SQLite is included with Python
# Uncomment for PostgreSQL:
# psycopg2-binary==2.9.7

# Uncomment for MySQL:
# mysqlclient==2.2.0

# Security
cryptography==41.0.7

# API documentation
drf-spectacular==0.26.5

# Image processing (if needed)
Pillow==10.1.0

# Timezone handling
pytz==2023.3

# YAML parsing
PyYAML==6.0.1

# File watching (development)
watchdog==3.0.0

# Backup and compression
python-magic==0.4.27

# System information
distro==1.8.0

# Network utilities
netifaces==0.11.0

# Configuration parsing
configparser==6.0.0

# Logging enhancements
structlog==23.2.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Testing (optional)
factory-boy==3.3.0
faker==19.12.1

# Performance monitoring (optional)
django-debug-toolbar==4.2.0

# Background tasks (optional)
celery==5.3.4

# File handling
pathlib2==2.3.7

# Validation
marshmallow==3.20.1

# CLI utilities
click==8.1.7

# Progress bars
tqdm==4.66.1

# Serialization
msgpack==1.0.7

# Text processing
chardet==5.2.0

# SSH/SFTP (if needed)
paramiko==3.3.1

# System commands
subprocess32==3.5.4; python_version < '3.0'

# Memory profiling (development)
memory-profiler==0.61.0

# Code quality (development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
pylint==3.0.2

# Testing framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Mock objects
responses==0.24.1

# Development server
django-extensions==3.2.3